feat: create component with routes
